+++ src/common/sourceUtils.ts
@@ -262,14 +262,6 @@ export async function checkContentHash(
   return result ? absolutePath : undefined;
 }
 
-export function positionToOffset(text: string, line: number, column: number): number {
-  let offset = 0;
-  const lines = text.split('\n');
-  for (let l = 1; l < line; ++l) offset += lines[l - 1].length + 1;
-  offset += column - 1;
-  return offset;
-}
-
 interface INotNullRange {
   line: number;
   column: number;