+++ src/ioc.ts
@@ -54,6 +54,7 @@ import { ILogger } from './common/logging';
 import { Logger } from './common/logging/logger';
 import { createMutableLaunchConfig, MutableLaunchConfig } from './common/mutableLaunchConfig';
 import { CodeSearchStrategy } from './common/sourceMaps/codeSearchStrategy';
+import { IRenameProvider, RenameProvider } from './common/sourceMaps/renameProvider';
 import { CachingSourceMapFactory, ISourceMapFactory } from './common/sourceMaps/sourceMapFactory';
 import { ISearchStrategy } from './common/sourceMaps/sourceMapRepository';
 import { ISourcePathResolver } from './common/sourcePathResolver';
@@ -357,7 +358,7 @@ export const provideLaunchParams = (
     .to(CachingSourceMapFactory)
     .inSingletonScope()
     .onActivation(trackDispose);
-
+  container.bind(IRenameProvider).to(RenameProvider).inSingletonScope();
   container.bind(DiagnosticToolSuggester).toSelf().inSingletonScope().onActivation(trackDispose);
 
   container.bind(BreakpointsPredictor).toSelf();